Add load directive to all modified files in previous commit
It's not needed in django 1.5, but if we want to have compatibility to
django 1.4, we should add it for now.

Changes made with few scripts:

```
for f in `git diff "HEAD" "HEAD~1" --name-only`
do
sed -i '/extends/ a\
{% load url from future %}
' $f
done
```

To add needed line after line with extends directive (because extends
directive has to be the first in the file).

However not all files have this directive, so we had to find the files
to which we didn't add the load directive:

```
for f in `git diff "HEAD" "HEAD~1" --name-only`
do
if grep "load url from future" > /dev/null $f
then
else; echo $f
fi
done > rest_of_the_files
```

When we had this list, the rest is to add the line in the beginning of
the file:

```
for f in `cat rest_of_the_files`
do
sed -i '1 i\
{% load url from future %}
' $f
done
```