Adds configuration for the linter
We are beginning to use the pep8speaks bot for our pull requests. The
bot is able to comment on the pull requests without configuration, but
is a bit too eager, he mentions all errors in the touched files. We
change it to only suggesting fixes on the touched lines (I hope).