Fix all is_student checks
The Student model is loading its corresponding User object  with
`select_related`, in order to limit database transactions.
Unfortunately, the reverse call (from user to student) ceases to fail
silently (return None) with an upgrade to Django 1.5. We are replacing
all unsafe calls to user.student with a proper check.