Partially added type annotations to models in Users app
After a brief email exchange we decided that adding type annotations specified in PEP-484 could be a good idea. It helps with linting and could be helpful if we decide to add MyPy to project in the future.
Work is not finished and it's continuation will be included in the next few commits.