Fixes problem with going back to courses (#523)
This commit fixes an issue with going back to the course page. When the course page was requested using AJAX request and we go back to it, the browser displays JSON instead of the rendered web-page. The issue originates in the fact, that the AJAX request and the course page use the same URL, and the browser caches the response.

The fix separates the URLs for the AJAX request and the HTML web-page.

Co-authored-by: BeataSzczuka <BeataSzczuka@users.noreply.github.com>